+++ ubo_app/services/050-rpi-connect/commands.py
@@ -26,7 +26,7 @@ from ubo_app.store.services.rpi_connect import (
     RPiConnectStatus,
     RPiConnectUpdateServiceStateAction,
 )
-from ubo_app.store.services.voice import ReadableInformation
+from ubo_app.store.services.speech_synthesis import ReadableInformation
 from ubo_app.utils.apt import is_package_installed
 from ubo_app.utils.async_ import create_task
 from ubo_app.utils.error_handlers import report_service_error